# code/trendsellr-domain/src/main/java/com/trendsellr/domain/model/product/Product.java
package com.trendsellr.domain.model.product;

import com.trendsellr.domain.model.Metadata;
import lombok.Builder;
import lombok.Data;

/**
 * Represents a product being tracked by the application.
 */

@Data
@Builder
public class Product {

    private String id;
    private String name;
    private String url;
    private String source;
    private Metadata metadata;
}